Add XML parsing API support
This introduces a new property on the options parameters called parsingMode. This can be set to auto (currently interpreted as html), html or xml.

Depending on parsingMode we should parse the document differently and use XmlDocument instead of HTMLDocument.

Currently the only difference is that jsdom will automatically determine the used parser by looking at the parsingMode instead of implying xml parsing with a <?xml directive and htmlparser2 as the configured parser.