Fix selenium startup on Windows
Seems like selenium outputs to stderr on windows. This might also have
been a recent change because of an update though.

Since we check for a specific line and can safely assume that selenium
won't output this line to stderr AND stdout we don't have to check if
the event was already emitted.