Use a real array for Node.prototype._childNodes
This change is a very big performance improvement because it avoids recalculating the "length" over and over.
Calling array methods on an real array is probably a lot faster too.
Node.prototype.childNodes still returns a live NodeList. This NodeList is lazily created to avoid defineProperty (which is very slow) during element construction